Make ObjFW compile on macOS/PowerPC 64
FossilOrigin-Name: 7b0cd744dc02d432e2a1a0bf8448b970ca6f05f833406dd9a5506b468eb6e101

+++ src/OFGZIPStream.m
@@ -249,7 +249,7 @@
 					    length: length];
 
 				_CRC32 = _OFCRC32(_CRC32, buffer, bytesRead);
-				_uncompressedSize += bytesRead;
+				_uncompressedSize += (uint32_t)bytesRead;
 
 				return bytesRead;
 			}
